requirements.txt
streamlit == 1.13.0
tensorflow == 2.8.0
google_api_python_client==1.12.8
numpy == 1.23.3
pandas == 1.4.2
tensorflow-datasets == 4.7.0
selenium == 4.7.2
webdriver_manager == 3.8.5
Pillow == 9.3.0
